Very custom module resolving ugliness

+++ commands/custom.vim
@@ -192,6 +192,43 @@ endfunction
 nnoremap <Leader>es :call EnableSpellCheck()<Enter>
 nnoremap <Leader>ds :call DisableSpellCheck()<Enter>
 
+" " Follow JavaScript references with proper babel (and module-resolver)
+" " resolution (requires a git repo in the project's root)
+" function FollowJsReference()
+"   " Find the reference and yank it into the r register
+"   exec "normal ^f'\"ryi'"
+"   let s:filename = @r
+"   if s:filename[0] == '.'
+"     " Treat it as a relative reference
+"     let s:file_with_js = s:filename . ".js"
+"     let s:file_with_index_js = s:filename . "/index.js"
+"     if filereadable(s:file_with_js)
+"       exec ":edit " . s:file_with_js
+"     elseif filereadable(s:file_with_index_js)
+"       exec ":edit " . s:file_with_index_js
+"     else
+"       echo "Couldn't find reference"
+"     endif
+"   else
+"     " Create a file at the root with the code for babel-node to resolve the
+"     " reference. Ideally, I'd do this with babel-node's eval, but it errors out.
+"     " Instead, I create a script, evaluate it then delete it.
+"     let s:script_content = 'console.log(require.resolve("' . @r . '"));'
+"     let s:root_dir = systemlist('dirname $(. ~/vim/vimrc/scripts/find-ancestor.sh .babelrc)')[0]
+"     let s:create_command = 'tee ' . s:root_dir . '/___resolve.js'
+"     call system(s:create_command, s:script_content)
+"     " Run babel-node from the root dir
+"     let s:babel_command = 'cd  ' . s:root_dir . '&& NODE_ENV=development babel ./___resolve.js | node'
+"     let s:out = system(s:babel_command)
+"     exec ":edit " . s:out
+"     " Clean up
+"     let s:delete_command = 'rm ' . s:root_dir . '/___resolve.js'
+"     call system(s:delete_command)
+"   endif
+" endfunction
+" nnoremap <Leader>gj :call FollowJsReference()<CR>
+
+
 " Follow JavaScript references with proper babel (and module-resolver)
 " resolution (requires a git repo in the project's root)
 function FollowJsReference()
@@ -210,20 +247,54 @@ function FollowJsReference()
       echo "Couldn't find reference"
     endif
   else
-    " Create a file at the root with the code for babel-node to resolve the
-    " reference. Ideally, I'd do this with babel-node's eval, but it errors out.
-    " Instead, I create a script, evaluate it then delete it.
-    let s:script_content = 'console.log(require.resolve("' . @r . '"));'
-    let s:root_dir = systemlist('dirname $(. ~/vim/vimrc/scripts/find-ancestor.sh .babelrc)')[0]
-    let s:create_command = 'tee ' . s:root_dir . '/___resolve.js'
-    call system(s:create_command, s:script_content)
-    " Run babel-node from the root dir
-    let s:babel_command = 'cd  ' . s:root_dir . '&& NODE_ENV=development babel ./___resolve.js | node'
-    let s:out = system(s:babel_command)
-    exec ":edit " . s:out
-    " Clean up
-    let s:delete_command = 'rm ' . s:root_dir . '/___resolve.js'
-    call system(s:delete_command)
+    if g:projectname == 'hive-web'
+      let s:out = expand(substitute(s:filename, '\~', '~/js-projects/hive-web/src', ''))
+      let s:file_with_js = s:out . ".js"
+      let s:file_with_index_js = s:out . "/index.js"
+      if filereadable(s:file_with_js)
+        exec ":edit " . s:file_with_js
+      elseif filereadable(s:file_with_index_js)
+        exec ":edit " . s:file_with_index_js
+      else
+        echom s:file_with_js
+        echom s:file_with_index_js
+        echo "Couldn't find reference"
+      endif
+    elseif g:projectname == 'sparrow'
+      let s:iscore = s:filename[:len('~/core') - 1] == '~/core'
+      if s:iscore
+        let s:out = expand(substitute(s:filename, '\~', '~/js-projects/sparrow/deps/lib', ''))
+      else
+        let s:out = expand(substitute(s:filename, '\~', '~/js-projects/sparrow/src', ''))
+      endif
+      let s:file_with_js = s:out . ".js"
+      let s:file_with_index_js = s:out . "/index.js"
+      if filereadable(s:file_with_js)
+        exec ":edit " . s:file_with_js
+      elseif filereadable(s:file_with_index_js)
+        exec ":edit " . s:file_with_index_js
+      else
+        echom s:file_with_js
+        echom s:file_with_index_js
+        echo "Couldn't find reference"
+      endif
+    else
+      " If it's not one of my projects, default to using babel
+      " Create a file at the root with the code for babel-node to resolve the
+      " reference. Ideally, I'd do this with babel-node's eval, but it errors out.
+      " Instead, I create a script, evaluate it then delete it.
+      let s:script_content = 'console.log(require.resolve("' . @r . '"));'
+      let s:root_dir = systemlist('dirname $(. ~/vim/vimrc/scripts/find-ancestor.sh .babelrc)')[0]
+      let s:create_command = 'tee ' . s:root_dir . '/___resolve.js'
+      call system(s:create_command, s:script_content)
+      " Run babel-node from the root dir
+      let s:babel_command = 'cd  ' . s:root_dir . '&& NODE_ENV=development babel ./___resolve.js | node'
+      let s:out = system(s:babel_command)
+      exec ":edit " . s:out
+      " Clean up
+      let s:delete_command = 'rm ' . s:root_dir . '/___resolve.js'
+      call system(s:delete_command)
+    endif
   endif
 endfunction
 nnoremap <Leader>gj :call FollowJsReference()<CR>
@@ -342,6 +413,8 @@ nnoremap <silent> ]d :on<CR>:next<CR>:Gdiff HEAD<CR>
 nnoremap <silent> [d :on<CR>:prev<CR>:Gdiff HEAD<CR>
 
 nnoremap <silent> <F9> :!adb shell input keyevent 82 && adb shell input keyevent 19 && adb shell input keyevent 23<CR><CR>
+nnoremap <silent> <F10> :!adb shell input keyevent 82<CR><CR>
+nnoremap <silent> <F11> :!adb reverse tcp:8081 tcp:8081<CR><CR>
 
 func! s:strfind(s,find,start)
   if type(a:find)==1
